+++ typings/custom.d.ts
@@ -6,8 +6,6 @@
 
 declare module 'autoprefixer';
 
-declare module 'lodash';
-
 declare module 'stringify-object' {
   export interface Options {
     indent?: string;